Implement parallel worker pool to parallelize directories, fix leaks and bugs
This implements deque similar to crossbeam in Rust, which uses the
Chase-Lev algorithm for a lock-free parallel data structure. This
is used to iterate over all directories and files in parallel.

It allocates on the heap, there is a TODO to see if this can be avoided.

Also fixes a bug in gitignore after parallelizing the work. Fixes leaks
that were appearing in the unit tests.